move getTokenInformation to contracts workspace
