+++ packages/frontend/src/utils/token.ts
@@ -1,5 +1,4 @@
-import { assert } from "@std/assert";
-import { Address, PublicClient, toBytes, zeroAddress } from "viem";
+import { Address, toBytes, zeroAddress } from "viem";
 import { abi, tokenAddresses } from "@massmarket/contracts";
 
 // Any utility functions for tokens
@@ -20,29 +19,4 @@ export function getTokenAddress(
     throw new Error(`Token not found for ${symbol} on chainId: ${chainId}`);
   }
   return toBytes(tokenAddress);
-}
-
-export function getTokenInformation(
-  publicClient: PublicClient,
-  tokenAddress: `0x${string}`,
-): Promise<[string, number]> {
-  assert(publicClient.chain, "publicClient.chain is undefined");
-  if (tokenAddress === zeroAddress) {
-    return new Promise((resolve) => {
-      resolve(["ETH", 18]);
-    });
-  }
-  const symbol = publicClient.readContract({
-    address: tokenAddress,
-    abi: abi.eddiesAbi,
-    functionName: "symbol",
-    args: [],
-  }) as Promise<string>;
-  const decimal = publicClient.readContract({
-    address: tokenAddress,
-    abi: abi.eddiesAbi,
-    functionName: "decimals",
-    args: [],
-  }) as Promise<number>;
-  return Promise.all([symbol, decimal]);
 }
